requirements.txt
boto3>=1.34.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
